refactor(web/services): create a service for technology order
